<fix> offer-댓글-직접 작성하기 redux 구현
